avoid hanging on output

+++ src/types.h
@@ -480,9 +480,9 @@ inline bool is_ok(Move m) {
 
 // key as stored in a transposition table entry
 class TTEKey {
-  uint16_t key;
+  uint32_t key;
  public:
-  TTEKey(const Key& k) : key(k.key >> 48) {
+ TTEKey(const Key& k) : key(k.key >> /*((8-sizeof(key))*8) */ 48 ) {
   }
   friend bool operator!=(const TTEKey& x, const TTEKey& y){
     return x.key != y.key;
